terceira do dia

+++ Workshop/App/ExcluirImagem.php
@@ -3,7 +3,8 @@
     $caminho = $_POST['caminho'];
     require 'Class/ImagemDataObject.php';
     $imagem = new ImagemDataObject;
-    $imagem::excluir($id);
-    unlink("../public/$caminho");
+    if($imagem::excluir($id)){
+        unlink("../public/$caminho");
+    }
     header('Location: ../public/galeria.php');
